analyzer/requirements.txt
python-dotenv==1.0.1
faster-whisper==0.10.0
openai==1.75.0
tiktoken==0.6.0
pydub==0.25.1